requirements.txt
praw==7.7.1
textblob==0.17.1
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
numpy==1.26.2